requirements.txt
requests
python-telegram-bot
wikipediaapi